added task prototype
